\name{smallcell}
\alias{smallcell}
\title{Smallcell Lung Cancer Data}
\description{
There are 121 observations on 4 variables.
Arm is the indication of two treatments. (either C + E or E + C).
Entry is the age of the patient at entry.
Survival is the survival time and 
indicator is the censoring indicator (right censoring).
For more details please see the reference below.

Data are from 
Ying, Z., Jung, SH, and Wei, LJ (1995). 
Median regression analysis with censored data.
Journal of the American Statistical Association,
90, 178-184.
}
\usage{data(smallcell)}
\format{A data frame containing 121 observations on 4 variables:
\tabular{rl}{
        [,1] \tab "arm"\cr
        [,2] \tab "entry"\cr
        [,3] \tab "survival"\cr
        [,4] \tab "indicator"}
}
\references{
Ying, Z., Jung, SH, and Wei, LJ
(1995). Median regression analysis with censored data.
\emph{Journal of the American Statistical Association}, 90, 178-184.

Maksymiuk, A. W., Jett, J. R., Earle, J. D., Su, J. Q., Diegert, F. A., Mailliard,
 J. A., Kardinal, C. G., Krook, J. E., Veeder, M. H., Wiesenfeld, M.,
 Tschetter, L. K., and Levitt, R. (1994). Sequencing and Schedule Effects
 of Cisplatin Plus Etoposide in Small Cell Lung Cancer Results of a North
 Central Cancer Treatment Group Randomized Clinical Trial. 
 \emph{Journal of Clinical Oncology}, 12, 70-76.
}
\keyword{datasets}
